Add commands Jump, Call and Return
This is the minimal change version.

There were 3 possible solution:

#1) Remove these commands

#2) Change the command buffers so the client side
can not see the "get" pointer and therefore force
the client side to use tokens to say in sync.

It's the cleaner solution. It lets you put commands
in any shared memory thereby getting rid of the ring
buffer commands. 

But, it required major changes. After going down
that path for an entire day I punted for...

#3) Implement the commands assuming there is only
one buffer the commands can be in (what is currently
called the ring_buffer). With the jump commands
available it no longer has to be used as a ring
buffer. A Pepper or NaCl app can use it however
they want but, unlike solution #2 they only get
the one buffer.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/555129

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37446 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 67e5077346cde830210611c2f00fb550d563fcd3

+++ gpu/command_buffer/service/cmd_parser.cc
@@ -46,7 +46,7 @@ parse_error::ParseError CommandParser::ProcessCommand() {
     return parse_error::kParseInvalidSize;
   }
 
-  if (static_cast<size_t>(header.size + get) > entry_count_) {
+  if (static_cast<int>(header.size) + get > entry_count_) {
     DLOG(INFO) << "Error: get offset out of bounds";
     return parse_error::kParseOutOfBounds;
   }
